refactor: uffd: move handler logic to own module
Rename handler.rs to main.rs, which now includes:
- parsing cli arguments
- listening on the socket
- parsing stream to get information sent from Firecracker
- getting Firecracker's credentials
- creating uffd handler object
- polling for and handling events

Create handler.rs to keep implementation of uffd handler:
- copying faulted pages from memory file
- zeroing out pages that have been removed
- updating memory mappings
